fix(watch): invalidate BM25/cache after watch-ingested papers
watch_runner.run_watch() ingested PDFs directly without calling the
route layer's cache/index invalidation, so watch-ingested papers were
invisible to search and served stale from the retrieval cache. Extract
the shared helper into cache_refresh.py and call it from watch_runner
after any paper is actually indexed.
